Iniciado Criação de Logs

+++ src/main/java/br/com/ibtechnology/cpsweb/support/user/UserMB.java
@@ -1,7 +1,11 @@
 package br.com.ibtechnology.cpsweb.support.user;
 
 import java.util.List;
+import java.util.ResourceBundle;
 
+import javax.faces.application.Application;
+import javax.faces.application.FacesMessage;
+import javax.faces.context.FacesContext;
 import javax.inject.Inject;
 import javax.inject.Named;
 
@@ -11,6 +15,7 @@ import org.springframework.context.annotation.Scope;
 import org.springframework.web.context.WebApplicationContext;
 
 import br.com.ibtechnology.cpsweb.model.entities.UserEntity;
+import br.com.ibtechnology.cpsweb.model.repositories.IGroupRepository;
 import br.com.ibtechnology.cpsweb.model.repositories.IUserRepository;
 import br.com.ibtechnology.cpsweb.util.BaseBeans;
 
@@ -27,6 +32,12 @@ public class UserMB extends BaseBeans {
 	@Inject
 	private IUserRepository userRepository;
 	
+	@Inject
+	private IGroupRepository groupRepository;
+	
+	@Inject
+	private FacesContext context;
+
 	private List<UserEntity> users;
 
 	private UserEntity selectedUser;
@@ -54,7 +65,13 @@ public class UserMB extends BaseBeans {
 
 	public void delete() {
 		if (this.selectedUser != null) {
+			if(!this.selectedUser.isProtected()){
 			this.userRepository.delete(this.selectedUser.getId());
+			}else
+			{
+				FacesMessage msg = new FacesMessage(this.getResourceProperty("labels", "user_cant_deleted"), this.selectedUser.getUsername());
+		        FacesContext.getCurrentInstance().addMessage(null, msg);
+			}
 		}
 	}
 
@@ -71,6 +88,15 @@ public class UserMB extends BaseBeans {
 			logger.error(e.getMessage(), e);
 		}
 	}
+	
+	public void testClass(){
+		if(this.selectedUser != null){
+		System.out.println(this.selectedUser.getUsername());
+		}else {
+			System.out.println("Usuário Nulo");
+			
+		}
+	}
 
 	public void unselectUser() {
 		this.selectedUser = null;
@@ -84,4 +110,11 @@ public class UserMB extends BaseBeans {
 		this.selectedUser = selectedUser;
 	}
 
+	private String getResourceProperty(String resource, String label) {
+		context = FacesContext.getCurrentInstance();
+		Application application = this.context.getApplication();
+		ResourceBundle bundle = application.getResourceBundle(this.context, resource);
+
+		return bundle.getString(label);
+	}
 }
